Skeleton Enemy Set Up.

+++ Assets/Scripts/Enemy/Enemy.cs
@@ -6,7 +6,7 @@ public abstract class Enemy : MonoBehaviour
 {
 
     [SerializeField] protected int health;
-    [SerializeField] protected int speed;
+    [SerializeField] protected float speed;
     [SerializeField] protected int gems;
     [SerializeField] protected Transform pointA, pointB;
